PydanticAI/requirements.txt
pydantic==2.11.0
pydantic-ai==0.0.46
openai==1.69.0
python-dotenv==1.1.0
pytest==8.3.5
nest_asyncio==1.6.0